Add files via upload
Split the input and output portions of the R script, and done some redesign too. This can make it much easier to change the output, as just one file needs to be changed and the recordings' R scripts will call it, as opposed to needing to change each recordings' R script individually.