Add files via upload
Too many changes to too many things to really comment, but I'll try
Made modular scripts for this stuff, so there are input R scripts that read in the data and set certain variables and control, and there an output R script for functions and graphs that are shared. This way it is much easier to have changes propagate.
The new input script for specific GPU analysis will also loop through the individual locations, so it will generate all of the files for the combined and individual data. The Python script for the GPU analysis will grab the same output script needed for individual datasets, so the output will have identical results.
The new scripts for the High data processing will now place specific GPU results (still just for high) in their folders, for easier sorting.
The Python scripts will now identify a Locations.txt file, so they can place in the location information for you, if the file is present, so you do not need to manually add it to each script. The Combined datasets will have the location stored in it, but for the individual datasets it is just kept in the R file.