Fixes to handling Extended Y-axis in Diff graphs
Two changes for handling the Consecutive Difference graphs when the Y-axis is extended.
The "EXT" filename label is now after the data name, for better alphabetical  sorting of the output files.
An example of using an extended Y-axis, though with the default value, has been added for easier use later.