overAPI and graphDIFF fixes
Fixed an issue with overAPI being hidden, so in situations it should be off it was hard to find. It is now next to testAPI.
Fixed an issue with graphDIFF as apparently coord_cartesian breaks the stat_density_2d layer I was using. The X and Y limits for that graph are now set in the scales, as they had been previously.
Additionally, there was a small inconsistency with the Y scale expansion between the versions of the script that I have corrected.