2020 Overhaul
These are the latest versions of the scripts following an overhaul. Many changes including to how controls are handled, and placing all manipulations to data in function environments.
The most significant changes are having R compress the combined CSVs and giving R the ability to search for and find the CSVs and TXTs itself, instead of relying on Python to do that work.