ding text wrapping for facet labels
By using label_wrap_gen the facet labels will now be wrapped when thy get too long. The length is controlled by "facWID" in the Input.r scripts. Location, API, Quality, and GPU labels are all set to be wrapped like this, though Location is the only one likely to need this