Update OCAT - Search - PA.r
The | to & is a change I forgot I made.
The change to CSV.config though is to make it possible to have all of the data written into a single, compressed CSV, if SUBSET is "All". Thanks to the GPU, API, Quality, and Location columns though, each run is unique. The purpose for this is an experiment I hope to try.