Scale Fill change
By setting drop to FALSE in scale_fill_hue the colors for the bars in the Means graph will be consistent regardless of the number of GPUs in the data. This setting means colors are assigned to the factor levels, even if the factor is not present in the data.