Change to CSV Reading
Sometimes I want to work from a CSV with all of the data and then filter it here. This new code will first check for the All.csv file and filter what it loads.
Have an example for filtering here that uses Quality, but can be easily changed or expanded on.