Fixed GPU coloring
The GPU fill color scale for the summary graph will now be created based on the number of GPUs in listGPU, rather than the number in the data. This will making the coloring consistent as long as the listGPU is consistent, regardless of which GPU is actually in the data.